# openapi.yaml
openapi: 3.1.1
info:
  title: Construgame api
  description: Description
  version: 1.0.0
servers:
  - url: https://api.construgame.com.br/v1.0
    description: prod
security:
  - accessToken: []
paths:
...
  /auth/web/token:
    post:
      summary: "Login web"
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: "Login succeeded"
          content:
            application/json:
              schema:
                required:
                  - accessToken
                  - refreshToken
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  userId:
                    type: string
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      nickname:
                        type: string
                      photo:
                        type: string
                      status:
                        type: string
                        enum: ["WAITING_CONFIRMATION", "ACTIVE"]
                      email:
                        type: string
                  roles:
                    type: array
                    items:
                      type: object
                      properties:
                        organizationId:
                          type: string
                        role:
                          type: string
                          enum: [owner, admin, player]
  /auth/app/token:
    post:
      summary: "Login app"
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - phone
              properties:
                phone:
                  type: string
                authCode:
                  type: string
                otpHash:
                  type: string
      responses:
        "200":
          description: "Login succeeded"
          content:
            application/json:
              schema:
                required:
                  - accessToken
                  - refreshToken
                  - userId
                  - roles
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
                  userId:
                    type: string
                  roles:
                    type: array
                    items:
                      type: object
                      properties:
                        organizationId:
                          type: string
                        role:
                          type: string
                          enum: [owner, admin, player]
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      name:
                        type: string
                      nickname:
                        type: string
                      photo:
                        type: string
                      status:
                        type: string
                        enum: ["WAITING_CONFIRMATION", "ACTIVE"]
                      email:
                        type: string
...
  /auth/web/password:
    post:
      summary: "Finish recover password"
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - userId
                - code
                - password
              properties:
                userId:
                  type: string
                code:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: "Password changed"
  /user:
    post:
      summary: "Create user"
      tags: [User]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
                - email
              allOf:
                - $ref: "#/components/schemas/userForm"
                - type: object
                  properties:
                    organization:
                      type: object
                      required:
                        - organizationId
                        - role
                      properties:
                        organizationId:
                          type: string
                        role:
                          type: string
                          enum: [admin, manager, player]
                        sectorId:
                          type: string
                          format: uuid
                        sector:
                          type: string
                          example: "alvenaria"
                        position:
                          type: string
                          example: "carpinteiro"
      responses:
        "201":
          description: "User created"
          content:
            application/json:
              schema:
                required:
                  - id
                  - name
                $ref: "#/components/schemas/user"
  /user/{userId}:
    get:
      summary: "Get user"
      tags: [User]
      parameters:
        - $ref: "#/components/parameters/userId"
      responses:
        "200":
          description: "user found"
          content:
            application/json:
              schema:
                required:
                  - id
                  - name
                $ref: "#/components/schemas/user"
    patch:
      summary: "Update user"
      tags: [User]
      parameters:
        - $ref: "#/components/parameters/userId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/userForm"

      responses:
        "200":
          description: "user updated"
          content:
            application/json:
              schema:
                required:
                  - id
                  - name
                $ref: "#/components/schemas/organization"
    put:
      summary: "Replace user"
      tags: [User]
      parameters:
        - $ref: "#/components/parameters/userId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
              $ref: "#/components/schemas/userForm"

      responses:
        "200":
          description: "user replaced"
          content:
            application/json:
              schema:
                required:
                  - id
                  - name
                $ref: "#/components/schemas/user"
    delete:
      summary: "delete user"
      tags: [User]
      parameters:
        - $ref: "#/components/parameters/userId"
      responses:
        "200":
          description: "user deleted"
...
  /user/{userId}/organization:
    get:
      summary: "List user organizations"
      tags: [Members]
      parameters:
        - $ref: "#/components/parameters/userId"
      responses:
        "200":
          description: "Organization found"
          content:
            application/json:
              schema:
                allOf:
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          required:
                            - id
                            - name
                          allOf:
                            - $ref: "#/components/schemas/organization"
                  #- $ref: "#/components/schemas/pagination"
...
components:
  parameters:
...
    userId:
      in: path
      name: userId
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    organizationForm:
      type: object
      properties:
        name:
          type: string
        ownerId:
          type: string
        photo:
          type: string
    organization:
      allOf:
        - type: object
          properties:
            id:
              type: string
        - $ref: "#/components/schemas/organizationForm"
    userForm:
      type: object
      properties:
        name:
          type: string
        nickname:
          type: string
        signedTermsOfUse:
          type: boolean
        email:
          type: string
        phone:
          type: string
        photo:
          type: string
    user:
      allOf:
        - type: object
          properties:
            id:
              type: string
            emailConfirmed:
              type: string
            phoneConfirmed:
              type: string
            status:
              type: string
              enum: ["WAITING_CONFIRMATION", "ACTIVE"]
        - $ref: "#/components/schemas/userForm"
...
    pagination:
      type: object
      properties:
        pagination:
          type: object
          properties:
            currentPage:
              type: number
            itemsPerPage:
              type: number
            # pagecount:
            #   type: number
            # itemCount:
            #   type: number
...
  securitySchemes:
    accessToken:
      type: http
      scheme: bearer